experiment with other "modes" for Ghosts mod
Inspired by some discussion on the Terrafusion discord server.

This uses skill selection as a mode switch.

* skill 0 = Same as previous mod behavior; i.e. no monsters.

* skill 1 = Monsters are present but not touchable/shootable, and are not aware of player (permanent "notarget").

Skill 1 is the interesting addition IMO.

But since there's another two skill levels to play with I also did:

* skill 2 = Monsters are present, not aware of player, touchable/shootable, but only have 1 health.

* skill 3 = Basically just permanent "notarget". Monsters have full health.

Also this change fixes what appears to be a bug with flying/swimming monsters, which were always being removed even if tourism mode (showturtle) was disabled.

Haven't yet played through the whole game in the new modes to verify that it is completable.